+++ CHANGELOG.md
@@ -16,6 +16,8 @@ for further information about branching and tagging conventions.
 - Show available base asset amount on sale creation form
 - Warning message when showing recovery seed on signup page
 - QR code plugin wrapper
+- Added error handler
+- Added error codes localization
 
 #### Changed
 - Modularized: